[nailgun] Fixed error with memory calculation

+++ nailgun/nailgun/api/handlers/node.py
@@ -86,7 +86,7 @@ class NodeCollectionHandler(JSONHandler):
         orm().add(node)
         orm().commit()
         try:
-            ram = str(round(int(node.meta['memory']) / 1073741824, 1))
+            ram = str(round(float(node.meta['memory']) / 1073741824, 1))
         except (KeyError, TypeError, ValueError):
             ram = "unknown"
         cores = str(node.meta.get('cpu', {}).get('total', "unknown"))
